+++ examples/QuadView_Basic/QuadView_Basic.java
@@ -78,7 +78,7 @@ public class QuadView_Basic extends PApplet {
     // display results
     background(0);
     for(PeasyCam cam : cameras){
-      float[] view = cam.getViewport();
+      int[] view = cam.getViewport();
       image(cam.getCanvas(), view[0], view[1], view[2], view[3]);
     }
   }
